backend/requirements.txt
google-generativeai
newspaper3k
Pillow
pytesseract
openai-whisper
moviepy
python-dotenv
streamlit 
langchain 
langchain-google-genai 
requests 
tavily-python
deep_translator
langdetect